update files path
